tests/submission/W3C/domhtml/HTMLModElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - MOD</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/attributes-common-to-ins-and-del-elements.html#dom-mod-cite" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<ins cite="/support/html/sample.html" datetime="January 1, 2002">The INS element is used to indicate that a section of a document had been inserted.</ins>
<br>
<del cite="/support/html/sample.html" datetime="January 2, 2002">The DEL element is used to indicate that a section of a document had been removed.</del>
</p>



<script>
var domtest = async_test("The cite IDL attribute must reflect the element's cite content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcite;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("del");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcite = testNode.cite;

      assert_uri_equals(document.location.protocol,null,document.location.host,"sample.html",null,null,null,true,vcite);

    domtest.done();
}));

</script>
<p>The cite IDL attribute must reflect the element's cite content attribute.</p>
<div id="log"></div>
</body></html>
